// src/api/models/JourneyInstanceResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Journey Orchestrator API
 * Platform service responsible for executing configurable journeys using a state machine model. The service is business-agnostic and manages journey definitions and journey instances driven by events. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { JourneyStatus } from './JourneyStatus';
import {
    JourneyStatusFromJSON,
    JourneyStatusFromJSONTyped,
    JourneyStatusToJSON,
    JourneyStatusToJSONTyped,
} from './JourneyStatus';

/**
 * 
 * @export
 * @interface JourneyInstanceResponse
 */
export interface JourneyInstanceResponse {
    /**
     * 
     * @type {string}
     * @memberof JourneyInstanceResponse
     */
    instanceId?: string;
    /**
     * 
     * @type {string}
     * @memberof JourneyInstanceResponse
     */
    journeyCode?: string;
    /**
     * 
     * @type {number}
     * @memberof JourneyInstanceResponse
     */
    version?: number;
    /**
     * 
     * @type {string}
     * @memberof JourneyInstanceResponse
     */
    currentState?: string;
    /**
     * 
     * @type {JourneyStatus}
     * @memberof JourneyInstanceResponse
     */
    status?: JourneyStatus;
    /**
     * 
     * @type {Date}
     * @memberof JourneyInstanceResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof JourneyInstanceResponse
     */
    updatedAt?: Date;
}



/**
 * Check if a given object implements the JourneyInstanceResponse interface.
 */
export function instanceOfJourneyInstanceResponse(value: object): value is JourneyInstanceResponse {
    return true;
}

export function JourneyInstanceResponseFromJSON(json: any): JourneyInstanceResponse {
    return JourneyInstanceResponseFromJSONTyped(json, false);
}

export function JourneyInstanceResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): JourneyInstanceResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'instanceId': json['instanceId'] == null ? undefined : json['instanceId'],
        'journeyCode': json['journeyCode'] == null ? undefined : json['journeyCode'],
        'version': json['version'] == null ? undefined : json['version'],
        'currentState': json['currentState'] == null ? undefined : json['currentState'],
        'status': json['status'] == null ? undefined : JourneyStatusFromJSON(json['status']),
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function JourneyInstanceResponseToJSON(json: any): JourneyInstanceResponse {
    return JourneyInstanceResponseToJSONTyped(json, false);
}

export function JourneyInstanceResponseToJSONTyped(value?: JourneyInstanceResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'instanceId': value['instanceId'],
        'journeyCode': value['journeyCode'],
        'version': value['version'],
        'currentState': value['currentState'],
        'status': JourneyStatusToJSON(value['status']),
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'] == null ? value['updatedAt'] : value['updatedAt'].toISOString(),
    };
}