// src/api/models/transition-history-event-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Journey Orchestrator API
 * Platform service responsible for executing configurable journeys using a state machine model. The service is business-agnostic and manages journey definitions and journey instances driven by events. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EventInfo } from './event-info';

export interface TransitionHistoryEventResponse {
    /**
     * Unique identifier for the history event
     */
    'id': string;
    /**
     * ID of the journey instance
     */
    'instanceId': string;
    /**
     * Previous state before transition
     */
    'fromState': string;
    /**
     * New state after transition
     */
    'toState': string;
    'event': EventInfo;
    /**
     * When the transition occurred (ISO 8601)
     */
    'timestamp': string;
    /**
     * Additional context information
     */
    'metadata'?: { [key: string]: any; };
    /**
     * Sequence number for ordering same-millisecond events
     */
    'sequenceNumber'?: number;
}